Sprint17/library/requirements.txt
asgiref==3.5.2
Django==4.1
psycopg2-binary
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
djangorestframework
django-ckeditor
Pillow